Add new version mongodb orm

+++ routes/upload.js
@@ -23,21 +23,97 @@ router.post('/upload-beers', upload.single('file'), (req, res) => {
     }
     let fileCsv = req.file.buffer.toString('utf8');
     let toInserts = [];
-    fileCsv.split('\n').map(line => {
-        columns = line.split(';');
-        if (columns.length === 6) {
-            let line = {
-                name: columns[0].trim(),
-                picture: columns[1],
-                brewery: columns[2],
-                alcohol: columns[3],
-                type: columns[4],
-                description: columns[5].trim()
+    //Let's get every beers first
+    Beer.find({ __deleted: { $exists: true } }).execute().then(beers => {
+        fileCsv.split('\n').map(line => {
+            columns = line.split(';');
+            if (columns.length === 6) {
+                //Compute type
+                let type = columns[4];
+                switch(type) {
+                    case 'Spéciales brunes':
+                        type = 'brown';
+                        break;
+                    case 'Spéciales blondes':
+                        type = 'blond';
+                        break;
+                    case "Bières de Noël":
+                        type = 'christmas';
+                        break;
+                    case "Bières d'abbaye":
+                        type = 'abbey';
+                        break;
+                    case 'Ales, Stouts et IPA':
+                        type = 'ipa';
+                        break;
+                    case 'Bières trappistes':
+                        type = 'trappist';
+                        break;
+                    case 'Pils':
+                        type = 'pils';
+                        break;
+                    case 'Bières fruitées':
+                        type = 'fruity';
+                        break;
+                    case 'Saisons':
+                        type = 'season';
+                        break;
+                    case 'Bières blanches':
+                        type = 'white';
+                        break;
+                    case 'Rouges des Flandres':
+                        type = 'fruity';
+                        break;
+                    case 'Lambics et gueuzes':
+                        type = 'lambic';
+                        break;
+                    case 'Spéciales ambrées':
+                        type = 'amber';
+                        break;
+                    case 'Bières thématiques':
+                        type = 'special';
+                        break;
+                }
+                let line = {
+                    name: columns[0].trim(),
+                    picture: columns[1],
+                    brewery: columns[2],
+                    alcohol: columns[3],
+                    type,
+                    description: columns[5].trim()
+                }
+                let lookalikeBeer = beers.find(beer => Utils.similar(line.name, beer.name) > 0.85);
+                if(lookalikeBeer) {
+                    lookalikeBeer.description = line.description;
+                    lookalikeBeer.save();
+                } else {
+                    toInserts.push(new Beer(line));
+                }
             }
-            toInserts.push(new Beer(line));
-        }
+        });
+        makeBeers(toInserts, (result) => onMakeBeers(res, result), false);
+
     });
-    makeBeers(toInserts, (result) => onMakeBeers(res, result));
+    //Mistergoodbeer beer type
+    //{ 'blond', 'white', 'amber', 'brown', 'black' }
+
+    //Bierbel beer type
+    /*
+          'Bières thématiques',
+            'Spéciales brunes',
+            'Spéciales blondes',
+            "Bières d'abbaye",
+            'Bières de Noël',
+            'Spéciales ambrées',
+            'Ales, Stouts et IPA',
+            'Bières trappistes',
+            'Pils',
+            'Bières fruitées',
+            'Saisons',
+            'Bières blanches',
+            'Rouges des Flandres',
+            'Lambics et gueuzes',
+    */
 });
 
 const onMakeBeers = (res, result) => {
@@ -45,12 +121,18 @@ const onMakeBeers = (res, result) => {
 }
 
 //Clean the collection and create a new one from csv
-const makeBeers = (beers, callback) => {
-    Beer.drop().then((result) => {
+const makeBeers = (beers, callback, toDelete=true) => {
+    if (toDelete) {
+        Beer.drop().then((result) => {
+            Beer.insertMany(beers).then(res => {
+                callback(res);
+            })
+        });
+    } else {
         Beer.insertMany(beers).then(res => {
             callback(res);
-        })
-    });
+        });
+    }
 }
 
 router.post('/upload-bars', upload.single('file'), (req, res) => {
@@ -246,7 +328,7 @@ router.post('/upload-mister', upload.single('file'), (req, res) => {
         let tags = bar.tags;
         let characteristics = bar.characteristics;
         let name = bar.name;
-        let keywords = bar.keywords;
+        let keywords = bar.keyword;
         let minPrice = bar.regularPrice;
         let minPriceHappy = bar.happyHourPrice;
         let location = {
